feat: #66 Suspense 추가
